Add sign out tab
The sign out tab was designed to be its separate page. However, there
didn't seem to be a way to add a new page through a theme. So, instead,
I created this solution where the tab click appears to open a new page,
but really just swaps the content of the current page.

This commit also pulled some updates to translations which had not been
added previously.