Javascript for email validation
New npm module added for validation. FTL cannot directly use modules, so the validation function is copied to js.

Validating the whole form has changed, because validity depends on email and checkbox.

If email is invalid, it is validated as user types new one. Email is not marked as invalid, while typing or first keypress would render error.
Checkbox is validated every time it is checked.

Added error message for the checkbox too.