(HP-940) Create a common email template
The template has a koros-pattern, logo+title, content area and footer

Recurring css styles are in theme.properties. Email template styles must be in element's style-attribute and classes should not be used to gain maximum compatibility with email clients.

The images are now stored in a cloud service, because right now Keycloak cannot pass url.resourcesPath to email templates, so urls referring to the server cannot be used.