Added translation validation
Translations can have HTML that is hard to validate by hand. Added a validator for it.

Placeholder keys, like "{0}", must have "{" with matching "}".

Also the order of the keys {1}...{N} is matched between translations. The order should be same. If not, a warning is shown.

HTML-elements should have same attribute order. Attribute values are not matched. If not, a warning is shown.

Untranslated keys raise warnings.

If validation fails, translations are not updated. The console will show an error.

Validation is considered a failure if HTML is invalid or curly braces do not have matching "{" and "}"

Updated the README.md. It has better explanation what the validation does.

Translation keys should have only chars suitable for an object key