HP-643 Use Gulp for building
Replace individual calls to npm-sass with Gulp.

The (quite old version of) npm-sass used the deprecated libsass library
for the SCSS-to-CSS transformation. The Gulp pipeline uses the Dart Sass
project. As a result the built CSS files look a bit different, but are
effectively the same. Changes concern whitespace, quoting style and
color value format.