HP-643 Calculate hashes for the CSS files
Use the gulp-rev-all plugin to calculate hashes from the CSS files'
contents. Normally it would also put the hash into the file name. This
is prevented by using a `transformFilename` function, which does not
modify the filename in any way. But as a result of this step in the
build pipeline, the Vinyl `file` object contains the hash which can then
be used later.