HP-1444 Dedicate login info template's default view to email verification
The info.ftl template is used for many purposes. There was already
special handling for VERIFY_EMAIL and UPDATE_PASSWORD required action
cases. In all other cases (the default view) a message provided by
Keycloak was shown. That message is a general "your account was
updated" kind of message.

The default case of this view is currently used for only one purpose:
acknowledging a successful email verification for the user. The default
view is thus changed to show a detailed message what the user achieved.