HP-1972 Fix "logout completed" view
In commit 0808cf474065b8ec281e30d47ffcf29d99c511cf the `info.ftl`
template's default branch was dedicated to the email verification
acknowledgement. The showing of any provided message(s) was removed. But
it turned out that showing of such message is used in the "logout
completed" case. So in this commit that functionality is restored, but
in a slightly different way.

The code used to show the required actions list is copied from
Keycloak's base theme. The code that was removed in commit
0808cf474065b8ec281e30d47ffcf29d99c511cf always showed a colon, whether
there were some required actions following it or not. The code in this
commit only adds that colon if there are required actions.

The return link to `client.baseUrl` isn't included. In the "logout
completed" case the `client` points to the Keycloak's built-in `account`
client and its base URL link points to something like `<Keycloak base
url>/auth/realms/helsinki-tunnistus/account/`. In the current Helsinki
tunnistus configuration the `account` client is disabled, so visiting
that URL would just produce an error.