HP-1017 Hide login methods that the current client doesn't support
The login methods that a client supports are listed in its
configuration. That configuration is now used to exclude parts of the
login form, showing only the relevant parts for the supported login
methods. If the client is missing this configuration, all login methods
are allowed.

This commit doesn't modify the styling of the elements. Only the logic
is included here.